% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{batt_totals}
\alias{batt_totals}
\title{Calculate Battery Totals}
\usage{
batt_totals(data, batt, value_to_find = "Selected", label_regex = "(?<=:).*$")
}
\arguments{
\item{data}{A data frame}

\item{batt}{A character string representing the battery identifier used to find relevant variables within the dataset.}

\item{value_to_find}{A character string representing the value to filter the totals by. Default is \code{"Selected"}.}

\item{label_regex}{A regular expression, used to extract labels from variable attributes. Default is \code{"(?<=:).*"} which extracts all text following a colon.}
}
\value{
A data frame containing the item labels and their corresponding total values.
}
\description{
This function calculates the totals for specific variables within a dataset, based on a given battery identifier. It extracts the relevant variables, computes totals, and filters them according to specified criteria.
}
\details{
Filters the results to include only those with a specified value.
It also extracts item labels based on a regular expression pattern.
}
\examples{
\dontrun{
# Example usage:
result <- batt_totals(data = my_data, batt = "Q23")
}

}
