% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{crosstab}
\alias{crosstab}
\title{Unweighted crosstab of one variable by another}
\usage{
crosstab(data, var, by, na.rm = FALSE, digits = 2, complete = TRUE)
}
\description{
The transpose of \code{totals_by}, which tends be more readable
}
\details{
For now, na.rm = FALSE is required for getting accurate distributions, since
the \code{n} column provides raw n for when missing values are included.
}
