% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{crosstab_list}
\alias{crosstab_list}
\title{Unweighted crosstab of multiple variables by one crossing variable}
\usage{
crosstab_list(data, vars, by, na.rm = FALSE, digits = 2, complete = TRUE)
}
\description{
Unweighted crosstab of multiple variables by one crossing variable
}
