% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_label}
\alias{find_label}
\title{Find variables in a dataset whose label attribute match some pattern}
\usage{
find_label(data, pattern)
}
\description{
Sometimes you don't know what a variable is called, but you do know
what its variable label might contain. This function lets you search
through a dataset and returns you variables and their factors which match
a regex.
}
