% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_level}
\alias{find_level}
\title{Find variables in a dataset with certain levels}
\usage{
find_level(data, pattern)
}
\description{
Sometimes you don't know what a variable is called, but you do know
what factor levels it has. This function lets you search through a dataset
and returns you variables and their factors which match a regex.
}
