% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_vars}
\alias{find_vars}
\title{Find Variables by Pattern}
\usage{
find_vars(data, pattern, exclude = "(_other)|(o$)|(FLAG)|(Flag)|(oe)")
}
\arguments{
\item{data}{A data frame from which variable names are to be extracted.}

\item{pattern}{A character string representing the pattern to match variable names against.}

\item{exclude}{A character string representing the pattern to exclude from the matched variable names. Default is \code{"(_other)|(o$)|(FLAG)|(Flag)|(oe)"}.}
}
\value{
A character vector of variable names that match the specified pattern and do not match the exclusion pattern.
}
\description{
Identifies and returns variable names from a dataset that match a specified pattern, while excluding certain patterns.
}
\examples{
\dontrun{
# Example usage:
vars <- find_vars(data = my_data, pattern = "Q23")
}

}
