% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{inform_var_count}
\alias{inform_var_count}
\title{Inform the user of the number of variables matched}
\usage{
inform_var_count(data)
}
\description{
Inform the user of the number of variables matched
}
\keyword{Internal}
