% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noodles-package.R
\docType{package}
\name{noodles-package}
\alias{noodles}
\alias{noodles-package}
\title{noodles: R Utiliies for Gartner}
\description{
A grabbag of useful functions and data structures useful for analyses of Gartner data assets The package is designed specifically for the Customer Service and Support PPR team but is broadly useful for anyone at Gartner hoping to do routine data analysis in R.
}
\author{
\strong{Maintainer}: J.J. Moncus \email{jj.moncus@gartner.com}

}
\keyword{internal}
