% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{spss_freq}
\alias{spss_freq}
\title{Frequency distribution of a variable, a la SPSS}
\usage{
spss_freq(data, var, digits = 2)
}
\description{
We don't allow the \code{na.rm} argument (as in, we force na.rm = TRUE)
because of the point of the table: to display "valid percentages" (borrowing
that language from SPSS) next to the raw n values generated by \code{table}. Raw
n for \code{NA} is still provided, but with no percentage attached.
}
\details{
Because \code{complete} is a \code{get_totals} parameter, it has no effect on which levels
get raw n values from \code{table}. Raw n's will be generated for all levels, unless
we add functionality for \code{complete} to affect both raw n's and percentages.
For now, we force \code{complete = TRUE} so that we get percentages 0-n values.
}
