% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals}
\alias{totals}
\title{Unweighted percentages for response options on a single variable}
\usage{
totals(data, var, na.rm = FALSE, digits = 2, complete = TRUE)
}
\description{
Unweighted percentages for response options on a single variable
}
