% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals_by}
\alias{totals_by}
\title{Unweighted percentages for response options on a single variable, split
by a second}
\usage{
totals_by(data, var, by, na.rm = FALSE, digits = 2, complete = TRUE)
}
\description{
Unweighted percentages for response options on a single variable, split
by a second
}
