% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals_list}
\alias{totals_list}
\title{Unweighted percentages for response options on multiple variables}
\usage{
totals_list(data, vars, na.rm = FALSE, digits = 2, complete = TRUE)
}
\description{
Unweighted percentages for response options on multiple variables
}
