% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{batt_totals}
\alias{batt_totals}
\title{Calculate Battery Totals}
\usage{
batt_totals(data, batt, label_regex = "(?<=:).*$", value_to_find = "Selected")
}
\arguments{
\item{data}{A data frame containing the dataset to be analyzed.}

\item{batt}{A character string representing the battery identifier used to find relevant variables within the dataset.}

\item{label_regex}{A regular expression pattern used to extract labels from variable attributes. Default is \code{"(?<=:).*"} which extracts text following a colon.}

\item{value_to_find}{A character string representing the value to filter the totals by. Default is \code{"Selected"}.}
}
\value{
A data frame containing the item labels and their corresponding total values.
}
\description{
This function calculates the totals for specific variables within a dataset, based on a given battery identifier. It extracts the relevant variables, computes totals, and filters them according to specified criteria.
}
\details{
The function identifies variables associated with the specified battery, calculates their totals, and filters the results to include only those with a specified value. It also extracts item labels based on a regular expression pattern. For specific batteries (Q23_new or Q29_new), items with a total value of zero are removed.
}
\examples{
\dontrun{
# Example usage:
result <- batt_totals(data = my_data, batt = "Q23")
}

}
