% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noodles-package.R
\docType{package}
\name{noodles-package}
\alias{noodles}
\alias{noodles-package}
\title{noodles: R Utiliies for Survey Work}
\description{
A grabbag of useful functions for survey data wrangling and analysis.
}
\author{
\strong{Maintainer}: J.J. Moncus \email{jj.moncus@gartner.com}

}
\keyword{internal}
